// src/types.ts
export interface Credentials {
  apiKey: string;
}

export interface AccessToken {
  /**
   * A JWT token which can be used to authrize against the other API end-points.
   * The format of the token follows the JWT standard format (see jwt.io for an example).
   * This is the token that should be sent in in the Authorization header when calling the other API end-points.
   */
  access_token: string;

  /**
   * The token type.
   */
  token_type: "string";

  /**
   * The validity time in seconds of the token
   */
  expires_in: 3600;
}
// ...
export interface Transaction {
  /**
   * Financial transactionIdd from mobile money manager.
   * Used to connect to the specific financial transaction made in the account
   */
  financialTransactionId: string;

  /**
   * External id provided in the creation of the requestToPay transaction
   */
  externalId: string;

  /**
   * Amount that will be debited from the payer account.
   */
  amount: string;

  /**
   * ISO4217 Currency
   */
  currency: string;

  /**
   * Party identifies a account holder in the wallet platform.
   * Party consists of two parameters, type and partyId.
   * Each type have its own validation of the partyId
   *   - MSISDN - Mobile Number validated according to ITU-T E.164. Validated with IsMSISDN
   *   - EMAIL - Validated to be a valid e-mail format. Validated with IsEmail
   *   - PARTY_CODE - UUID of the party. Validated with IsUuid
   */
  payer: Payer;

  /**
   * Message that will be written in the payer transaction history message field.
   */
  payerMessage: string;

  /**
   * Message that will be written in the payee transaction history note field.
   */
  payeeNote: string;

  reason: TransactionFailureReason;

  status: TransactionStatus;
}

/**
 * The available balance of the account
 */
export interface AccountBalance {
  /**
   * The available balance of the account
   */
  availableBalance: string;

  /**
   * ISO4217 Currency
   */
  currency: string;
}

export interface Payer {
  partyIdType: PartyIdType;
  partyId: string;
}

export type PartyIdType = "MSISDN" | "EMAIL" | "PARTY_CODE";

export interface TransactionFailureReason {
  type: TransactionFailureReasonType;
  message: string;
}

export enum TransactionFailureReasonType {
  payeeNotFound = "PAYEE_NOT_FOUND",
  payerNotFound = "PAYER_NOT_FOUND",
  notAllowed = "NOT_ALLOWED",
  notAllowedTargetEnviroment = "NOT_ALLOWED_TARGET_ENVIRONMENT",
  invalidCallbackUrlHost = "INVALID_CALLBACK_URL_HOST",
  invalidCurrency = "INVALID_CURRENCY",
  serviceUnavailable = "SERVICE_UNAVAILABLE",
  internalProcessingError = "INTERNAL_PROCESSING_ERROR",
  notEnoughFunds = "NOT_ENOUGH_FUNDS",
  payerLimitReached = "PAYER_LIMIT_REACHED",
  payeeNotAllowedToReceive = "PAYEE_NOT_ALLOWED_TO_RECEIVE",
  paymentNotApproved = "PAYMENT_NOT_APPROVED",
  resourceNotFound = "RESOURCE_NOT_FOUND",
  approvalReject = "APPROVAL_REJECTED",
  expired = "EXPIRED",
  transactionCancelled = "TRANSACTION_CANCELED",
  resourceAlreadyExist = "RESOURCE_ALREADY_EXIST"
}

// TODO: Get all valid transaction statuses
export type TransactionStatus = "SUCCESSFUL" | "PENDING" | "FAILED";